+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/SoraMediaChannel.kt
@@ -442,7 +442,7 @@ class SoraMediaChannel
          * JavaAudioDeviceModule の録音処理を一時停止または再開する
          * 内部で生成した AudioDeviceModule のみ制御可能であり、外部から差し込んだ AudioDeviceModule は対象外
          */
-        suspend fun setAudioRecordingPausedAsync(paused: Boolean): Boolean = peer?.setAudioRecordingPausedAsync(paused) ?: false
+        suspend fun setAudioRecordingPaused(paused: Boolean): Boolean = peer?.setAudioRecordingPaused(paused) ?: false
 
         /**
          * 録音停止中かどうかを返す